Redesigned API for DAP methods
Now the DAP interface better differentiate in names
utility (private) methods - starting with `_`, from
public methods that has to be reimplemented by sub-classes.
This should make the API a little bit less awkward.
Moreover, now `deep_learning_dap` forces the settings of
`to_categorical` to True - which is True in the 99% of
the cases.